feat: move unlink template provider to `shared` dir
