+++ lib/camunda-cloud/Modeler.js
@@ -13,6 +13,7 @@ import {
 } from 'bpmn-js-properties-panel';
 
 import replaceModule from './features/replace';
+import sharedModules from '../shared';
 
 import { commonModdleExtensions, commonModules } from './util/commonModules';
 
@@ -46,7 +47,8 @@ Modeler.prototype._camundaCloudModules = [
   rulesModule,
   zeebePropertiesProviderModule,
   cloudElementTemplatesPropertiesProvider,
-  replaceModule
+  replaceModule,
+  sharedModules
 ];
 
 Modeler.prototype._modules = [].concat(
