+++ lib/camunda-cloud/features/replace/ElementTemplatesReplaceProvider.js
@@ -1,15 +1,8 @@
 import {
-  isDifferentType
-} from 'bpmn-js/lib/features/popup-menu/util/TypeUtil';
-
-import {
   getBusinessObject,
   isAny
 } from 'bpmn-js/lib/util/ModelUtil';
 
-import {
-  getReplaceOptionGroups
-} from './ReplaceOptionsUtil';
 
 /**
  * A replace menu provider that allows to replace elements with
@@ -50,12 +43,6 @@ ElementTemplatesReplaceProvider.prototype.getPopupMenuEntries = function(element
     // convert our entries into something sortable
     let entrySet = Object.entries(entries);
 
-    if (this._elementTemplates.get(element)) {
-
-      // add unlink template option
-      this._addPlainElementEntry(element, entrySet);
-    }
-
     // add template entries
     entrySet = [ ...entrySet, ...this.getTemplateEntries(element) ];
 
@@ -66,85 +53,6 @@ ElementTemplatesReplaceProvider.prototype.getPopupMenuEntries = function(element
       return entries;
     }, {});
   };
-};
-
-/**
- * Adds the option to replace with plain element (unlink template).
- *
- * @param {djs.model.Base} element
- * @param {Array<Object>} entries
- */
-ElementTemplatesReplaceProvider.prototype._addPlainElementEntry = function(element, entries) {
-
-  const replaceOption = this._getPlainEntry(element, entries);
-
-  if (!replaceOption) {
-    return;
-  }
-
-  const [
-    insertIndex,
-    entry
-  ] = replaceOption;
-
-  // insert unlink entry
-  entries.splice(insertIndex, 0, [ entry.id, entry ]);
-};
-
-/**
- * Returns the option to replace with plain element and the index where it should be inserted.
- *
- * @param {djs.model.Base} element
- * @param {Array<Object>} entries
- *
- * @returns {Array<Object, number>}
- */
-ElementTemplatesReplaceProvider.prototype._getPlainEntry = function(element, entries) {
-
-  const {
-    options,
-    option,
-    optionIndex
-  } = findReplaceOptions(element) || { };
-
-  if (!options) {
-    return null;
-  }
-
-  const entry = {
-    id: 'replace-unlink-element-template',
-    action: () => {
-      this._elementTemplates.applyTemplate(element, null);
-    },
-    label: this._translate(option.label),
-    className: option.className
-  };
-
-  // insert after previous option, if it exists
-  const previousIndex = getOptionIndex(options, optionIndex - 1, entries);
-
-  if (previousIndex) {
-    return [
-      previousIndex + 1,
-      entry
-    ];
-  }
-
-  // insert before next option, if it exists
-  const nextIndex = getOptionIndex(options, optionIndex + 1, entries);
-
-  if (nextIndex) {
-    return [
-      nextIndex,
-      entry
-    ];
-  }
-
-  // fallback to insert at start
-  return [
-    0,
-    entry
-  ];
 };
 
 /**
@@ -207,45 +115,4 @@ export function isTemplateApplied(element, template) {
   }
 
   return false;
-}
-
-function getOptionIndex(options, index, entries) {
-  const option = options[index];
-
-  if (!option) {
-    return false;
-  }
-
-  return entries.findIndex(
-    ([ key ]) => key === option.actionName
-  );
-}
-
-/**
- * @param {ModdleElement} element
- *
- * @return { { options: Array<any>, option: any, optionIndex: number } | null }
- */
-function findReplaceOptions(element) {
-
-  const isSameType = (element, option) => option.target && !isDifferentType(element)(option);
-
-  return getReplaceOptionGroups().reduce((result, options) => {
-
-    if (result) {
-      return result;
-    }
-
-    const optionIndex = options.findIndex(option => isSameType(element, option));
-
-    if (optionIndex === -1) {
-      return;
-    }
-
-    return {
-      options,
-      option: options[optionIndex],
-      optionIndex
-    };
-  }, null);
 }
