+++ lib/shared/index.js
@@ -1,0 +1,8 @@
+import UnlinkTemplateReplaceProvider from './UnlinkTemplateReplaceProvider';
+
+export default {
+  __init__: [
+    'unlinkTemplateReplaceProvider'
+  ],
+  unlinkTemplateReplaceProvider: [ 'type', UnlinkTemplateReplaceProvider ]
+};
