+++ lib/shared/util/ReplaceOptionsUtil.js
